api/requirements.txt
PyJWT==2.1.0 